requirements.txt
gpustat
tensorboardX
wandb
torchsummary